[PLFM-9517] Move search query schemas to generic search namespace (#4)
* [PLFM-9517] Move search query schemas to generic search namespace

Move generic query/response schemas from table.search to search namespace
for reuse across both table-based and entity-based search. Table-specific
schemas (SearchIndex, SearchIndexState) move to search.table namespace.

New schemas:
- search/SearchQuery (removed searchIndexId, booleanFilters; added as AsyncRequestBody)
- search/SearchQueryResults (renamed from SearchResults; AsyncResponseBody)
- search/SearchIndexQuery (composition wrapper with searchIndexId + searchQuery)
- search/FacetRequest (added sortField + sortDirection)
- search/KeyRange (copied from search.query for clean break)

* [PLFM-9517] Update SearchQueryController types for new search namespace

Update stub controller to use SearchIndexQuery and SearchQueryResults
from the refactored search/search.table namespaces.

* [PLFM-9517] Move TextAnalyzer and ColumnAnalyzerOverride schemas to search.table namespace

Move TextAnalyzer and ColumnAnalyzerOverride JSON schemas from
table.search to search.table namespace. Update all Java imports
across DAOs, managers, services, controllers, and tests.

* Updating facet sort to be an enum

* [PLFM-9517] Fix client imports for search.table namespace move

Update SynapseClient and SynapseClientImpl imports from table.search to
search.table to match the schema relocation in f9fff3ce49.

* [PLFM-9517] Rename from/size to offset/limit and replace map fields with SearchFieldValue objects

Addresses code review feedback:
- Rename pagination fields from/size to offset/limit in SearchQuery and SearchQueryResults for API consistency
- Replace map-type fields/highlights in SearchHit with List<SearchFieldValue> (name/value pairs)
- Remove SearchIndex.json and SearchIndexState.json from this PR (deferred to entity-creation PR)
