+++ lib/lib-auto-generated/src/main/resources/schema/org/sagebionetworks/repo/model/search/SearchQueryType.json
