+++ lib/lib-auto-generated/src/main/resources/schema/org/sagebionetworks/repo/model/search/SortField.json
@@ -7,7 +7,7 @@
 		},
 		"direction": {
 			"description": "Sort direction.",
-			"$ref": "org.sagebionetworks.repo.model.table.search.SortDirection"
+			"$ref": "org.sagebionetworks.repo.model.search.SortDirection"
 		}
 	}
 }
