+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/search/TextAnalyzerManager.java
@@ -1,9 +1,9 @@
 package org.sagebionetworks.repo.manager.search;
 
 import org.sagebionetworks.repo.model.UserInfo;
-import org.sagebionetworks.repo.model.table.search.ListTextAnalyzersRequest;
-import org.sagebionetworks.repo.model.table.search.ListTextAnalyzersResponse;
-import org.sagebionetworks.repo.model.table.search.TextAnalyzer;
+import org.sagebionetworks.repo.model.search.table.ListTextAnalyzersRequest;
+import org.sagebionetworks.repo.model.search.table.ListTextAnalyzersResponse;
+import org.sagebionetworks.repo.model.search.table.TextAnalyzer;
 
 public interface TextAnalyzerManager {
